// frontend/src/app/core/models/admin.ts
export interface Admin{
    email: string;
    password: string;
    firstName:string;
    middleName:string;
    paternalLastName:string;
    maternalLastName:string;
    age: number;
    dni:string;
    phone:string;
    birthDate: string;
    profile: string;
    gender: string;
    nationality: string;

}